+++ CHANGELOG.md
@@ -7,8 +7,12 @@
 ### Added
 - Inject the `generator` attribute of the `@Id` annotation into the Java class name
 - The load parameter of the @OneToOne and @OneToMany annotations is injected as SQL, with support for syntax highlighting.
+- `*-translate.xml` files support syntax highlighting.
 - `@Id`注解的`generator`参数注入为java类名, 方便引用及跳转
 - `@OneToOne`和`@OneToMany`注解的`load`参数注入为sql，支持语法高亮
+- `*-translate.xml`文件支持语法注入
+- `*-translate.xml`文件中的`datasource`属性、`service`属性直接语法注入为`spring bean name`方便跳转（有可能不生效，很诡异）
+- `*-translate.xml` file use sqltoy icon
 
 ### Fixed
 - 修复`2024.1`中语法注入报错的问题